Moved DialogControllerImpl (PlatformOPtionPage.kt) => DialogControllerPane (Dialog.kt)
